修改部分bug 增加录音开始和结束提示音 修改选择好友页面ui

+++ code/src/com/rcplatform/phototalk/TakePhotoActivity.java
@@ -139,13 +139,16 @@ public class TakePhotoActivity extends Activity {
 	}
 
 	public void deleteFile(File file) {
-		if (file.isDirectory()) {
-			File[] files = file.listFiles();
-			for (File file2 : files) {
-				deleteFile(file2);
+		if (file.exists()) {
+			if (file.isDirectory()) {
+				File[] files = file.listFiles();
+				for (File file2 : files) {
+					deleteFile(file2);
+				}
+			} else {
+				file.delete();
 			}
 		}else{
-			file.delete();
 		}
 	}
 }
